services/data-collection/scraper/requirements.txt
# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
lxml==4.9.3
html5lib==1.1
webdriver-manager==4.0.1
playwright==1.40.0

# API and web server
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
starlette==0.27.0

# Database
pymongo==4.6.0
motor==3.3.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Storage
boto3==1.28.64
minio==7.2.0

# Data processing
pandas==2.1.2
numpy==1.26.1
pyarrow==14.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
tenacity==8.2.3
schedule==1.2.1
httpx==0.25.1
aiohttp==3.8.6

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.18.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
